Separate index and app
